requirements.txt
lpips==0.1.4
numpy==1.26.4
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.3.0
piq==0.8.0
pytorch_lightning==2.2.5
skimage==0.0
timm==1.0.3
torch==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
